add JSDoc to each process function
