+++ packages/babel-plugin/src/processK.ts
@@ -1,6 +1,15 @@
 import { NodePath, Node, types, template } from "@babel/core";
 import { Program } from "@babel/types";
 
+/**
+ * Processes the CallExpression nodes in the AST and replaces the 'k' function from '@kuma-ui/core'
+ * with the corresponding HTML tag. This allows usage of the 'k' function as a shorthand for creating
+ * styled components, e.g. <k.div> instead of <div>.
+ *
+ * @param {NodePath<Program>} node - The NodePath object representing the Program node.
+ * @param {typeof types} t - The Babel types object.
+ * @param {Record<string, string>} importedStyleFunctions - An object containing the imported styled functions.
+ */
 export const processK = (
   node: NodePath<Program>,
   t: typeof types,
