+++ packages/babel-plugin/src/visitor.ts
@@ -11,6 +11,7 @@ import { processHTMLTag } from "./processHTMLTag";
 import { Node } from "@babel/core";
 import { collectImportedStyled } from "./collectImportedStyled";
 import { processK } from "./processK";
+import { processTaggedTemplateExpression } from "./processTaggedTemplateExpression";
 
 export const styledFunctionsMap = new Map<string, Node[]>();
 
@@ -20,49 +21,49 @@ export const visitor = ({ types: t, template }: Core) => {
   let importedStyleFunctions: Record<string, string> = {};
 
   const visitor: PluginObj<PluginPass>["visitor"] = {
-    TaggedTemplateExpression(path) {
-      // Check if the tag is a CallExpression with the callee named 'styled'
-      const { node } = path;
+    // TaggedTemplateExpression(path) {
+    //   // Check if the tag is a CallExpression with the callee named 'styled'
+    //   const { node } = path;
 
-      const hasStyled = Object.keys(importedStyleFunctions).some(
-        (key) =>
-          t.isCallExpression(node.tag) &&
-          t.isIdentifier(node.tag.callee) &&
-          importedStyleFunctions[key] === node.tag.callee.name
-      );
-      if (!(t.isCallExpression(node.tag) && hasStyled)) return;
-      const componentArg = node.tag.arguments[0];
-      const cssStrings = node.quasi.quasis.map((quasi) => quasi.value.raw);
-      // Remove newlines and extra spaces from cssStrings, and concatenate them
-      const cssString = cssStrings
-        .map((str) => str.replace(/\s+/g, " ").trim())
-        .join("");
-      const className = !!cssString ? sheet.addRule(cssString) : undefined;
+    //   const hasStyled = Object.keys(importedStyleFunctions).some(
+    //     (key) =>
+    //       t.isCallExpression(node.tag) &&
+    //       t.isIdentifier(node.tag.callee) &&
+    //       importedStyleFunctions[key] === node.tag.callee.name
+    //   );
+    //   if (!(t.isCallExpression(node.tag) && hasStyled)) return;
+    //   const componentArg = node.tag.arguments[0];
+    //   const cssStrings = node.quasi.quasis.map((quasi) => quasi.value.raw);
+    //   // Remove newlines and extra spaces from cssStrings, and concatenate them
+    //   const cssString = cssStrings
+    //     .map((str) => str.replace(/\s+/g, " ").trim())
+    //     .join("");
+    //   const className = !!cssString ? sheet.addRule(cssString) : undefined;
 
-      const component = t.isStringLiteral(componentArg)
-        ? componentArg.value
-        : t.isJSXElement(componentArg)
-        ? componentArg
-        : "div";
-      const createElementAst = template.expression.ast(
-        `
-              (props) => {
-                const existingClassName = props.className || "";
-                const newClassName = "${className || ""}";
-                const combinedClassName = [existingClassName, newClassName].filter(Boolean).join(" ");
-                return React.${
-                  typeof component === "string"
-                    ? `createElement("${component}"`
-                    : `cloneElement(${component}`
-                }, {
-                  "data-kuma-ui": "${true}",
-                  ...props,
-                  className: combinedClassName,
-                });
-              }`
-      );
-      path.replaceWith(createElementAst);
-    },
+    //   const component = t.isStringLiteral(componentArg)
+    //     ? componentArg.value
+    //     : t.isJSXElement(componentArg)
+    //     ? componentArg
+    //     : "div";
+    //   const createElementAst = template.expression.ast(
+    //     `
+    //           (props) => {
+    //             const existingClassName = props.className || "";
+    //             const newClassName = "${className || ""}";
+    //             const combinedClassName = [existingClassName, newClassName].filter(Boolean).join(" ");
+    //             return React.${
+    //               typeof component === "string"
+    //                 ? `createElement("${component}"`
+    //                 : `cloneElement(${component}`
+    //             }, {
+    //               "data-kuma-ui": "${true}",
+    //               ...props,
+    //               className: combinedClassName,
+    //             });
+    //           }`
+    //   );
+    //   path.replaceWith(createElementAst);
+    // },
     // JSXElement(path: NodePath<JSXElement>) {
     //   const openingElement = path.get("openingElement");
 
@@ -85,10 +86,19 @@ export const visitor = ({ types: t, template }: Core) => {
     },
     Program: {
       enter(path) {
+        // Ensure that 'React' is imported in the file
         ensureReactImport(path, t);
         // Reset the importedStyleFunctions
         importedStyleFunctions = collectImportedStyled(path, t);
+        // Replace the 'k' function from '@kuma-ui/core' with the corresponding HTML tag
         processK(path, t, importedStyleFunctions);
+        // Process TaggedTemplateExpressions with styled components and generate the hashed classNames
+        processTaggedTemplateExpression(
+          path,
+          t,
+          template,
+          importedStyleFunctions
+        );
       },
     },
   };
